[FIX] helpdesk: open dashboard when no teams has sla_policy
Current behaviour:
If all the teams on the helpdesk app have sla_policy disabled,
you have a stacktrace when trying to open the helpdesk app, if a
ticket has been closed in the past 7 days.

Expected behaviour:
Should be able to open the helpdesk dashboard, without having to
check the sla_policy on one of the team.

Steps to reproduce:
- Install Helpdesk
- Remove the SLA Policies from all the teams in the Helpdesk
- Try reopening the Helpdesk app, you get a stacktrace.

Reason for the problem:
Key error on a result of a `read_group` of `sla_reached` and
`sla_reached_late`, those are not present in the result if the
option is not present on the teams.

Fix:
Replace the dictionary indexing (`['key']`) with a `get('key', False)`.

Affected versions:
- saas-15.2
- 16.0
- saas-16.1
- master

opw-3169337

closes odoo/enterprise#37467

X-original-commit: fe1ef4f3a1fa5f795066c241fd1ff5aeb29a9fb3
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>