[FIX] hr_payroll: todo_list use component Record
This commit aims to fix two bugs in the client action "hr_payroll_dashboard":
1. Opening the debug menu displays an error
2. Editing a note causes a crash

Why is this?
1. The useModel hook was created to be used only in a view. Using it
in a client action will cause the debug menu to think it is in a view,
which will cause the crash.

2. Using the html field in web_editor mode requires passing the setDirty
props to it. This will be used when the field is edited. In this case,
we use the FieldHtml component without this prop. So we will have a
crash when we edit the field.

Solution:
1. In this case, the only part of the component that needs a model
is the one that uses the web_editor html field. It is not necessary to
have a model to manage the addition/removal and editing of the name
of the notes. For the display of the html field, we will use the Record
component which will generate a template and display the field.

2. The use of the field Field will take care to add the props setDirty.

closes odoo/enterprise#37123

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>