[FIX] helpdesk: unread message invisible when notif in Odoo
Actual Behavior
===============

When handling notifications by email, the filter "Unread message" is
still available.

Expected Behavior
=================

we want the same behavior as seen in the project: The filter is only
available when we handle the message with Odoo

How to reproduce
===============

1. Install helpdesk
2. Set your notification handling method to email
3. Goto helpdesk > Tickets > All / My tickets
4. If you can on the filter menu, the "Unread Message" is visible

If we try to reproduce this manipulation the Project module, the "Unread
Message" is not available on the filter menu.

opw-3184919

closes odoo/enterprise#37128

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>