[FIX] appointment: avoid creating appointment link inside existing link
When a link is created via the "/Appointment" or "/Calendar" commands
inside an existing link (e.g. a button in website), we get nested anchor
tags in the DOM as result.

This commit avoids such result by changing the existing link's href and
content instead of creating a new one in such case.

This commit also fixes the missing call to `_t` for user-facing strings and
adds the missing related unit tests.

task-3130947
opw-3096108

closes odoo/enterprise#36839

Related: odoo/odoo#112904
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>