[FIX] l10n_be_reports: Partner VAT listing: fix the report when making an invoice with a 0% tax
Creating an invoice with a base amount superior to the threshold (250€) and a 0% tax on it crashed the report. This was due to missing COALESCEs in the query it ran : SUM() gives NULL as its result when called on empty data.

Once this was fixed, another COALESCE had to be added on the grouping key, otherwise the name of the partner would be displayed as "Unknown". This was because multiple  columns were added to the result with the name 'grouping_key', and not all of them contained a value: only the last one was retrieved when doing a get('grouping_key') on the result dict, giving None.

closes odoo/enterprise#37228

X-original-commit: e2dd37b95d979d96045656c2bc0c54209bfc06a9
Signed-off-by: John Laterre (jol) <jol@odoo.com>