[IMP] knowledge: adding clone button
This commit is adding a new button to the article view for users that can write
on specific articles.
We are adding the possibility for the user to clone an existing article if he
has write access on it.

By cloning an article the clone keeps the same parent, the same property
values, the item status and many more in addition of the fields copied
when using the `action_make_private_copy` method.
This is useful when a user wants to create many articles that are similar:
he creates the first one and then creates one or more clones that he slightly
modifies corresponding to its needs.

task-3186202

closes odoo/enterprise#37287

Signed-off-by: David Beguin (dbe) <dbe@odoo.com>