[IMP] knowledge: merge /index and /article_structure into a single command
This commit removes article structure command and makes the index command
show all children articles by default. A switch button has been added to allow
the user to switch between the 2 modes:
- Show all: show all the children articles
- Show Less: show only the direct children.

This does not impact existing blocks as only the command to create this block is
modified and still generates the same content. No migration are needed.

Task-3188750

Part-of: odoo/enterprise#37299