[REF] *: register widget descriptors instead of components
Before this commit, the widget's description was stored on the
component and this component was then registered.

Now, an object describing the widget is used on registration the same way
as it is done for views since b828cfc.
This split the component's description (props, template, ...) of
the widget's description ( component, extractProps, ...) and makes
it clearer.

We did the same thing for fields in 9f4622492c5cd84847368d32bf02b0b538a6dd69

Part of task: 3179751

closes odoo/enterprise#37215

Related: odoo/odoo#112962
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>