[FIX] website_delivery_ups: disable checkout when COD it not with UPS
Current behaviour:
The COD (Cash on Delivery) payment method is used only for UPS normally,
but you are able to checkout an order with another delivery method.

Expected behaviour:
You shouldn't be able to checkout with COD if the delivery method is
not UPS.

Steps to reproduce:
- Install Ecommerce, UPS Delivery
- Activate/Publish the COD payment provider and UPS delivery
- Add a product on the e-shop and checkout
- Select COD and another delivery method than UPS, you can checkout,
  when you shouldn't be able to.

Reason for the problem:
We don't handle the case in general.

Fix:
In the front-end, remove the COD payment option if we haven't selected
an UPS delivery method with `cash_on_delivery` activated.

Affected versions:
- 16.0
- saas-16.1
- master

opw-3130192

closes odoo/enterprise#37218

Signed-off-by: Piryns Victor (pivi) <pivi@odoo.com>