[REF] (website_)sale_(stock_)renting: single rental period
Before this commit, the period of rental was defined on each line of a
sale order. This allows more flexibility but comes along a complexity
restraining to develop further (as a deep integration with the stock
logic).

After this commit, the rental period will be defined on the sale order,
hence implying all listed rented products to be rented on the same
dates. The same applies to the cart on eCommerce.

This is the first step to improve the integration of the rental
app with other flows (main objective being an advanced integration
with the inventory management).

Now that the rental period is stored on the order, some disabled
flows for rental products are now enabled, including:

* quick add of the product to the ecommerce cart
* product catalog

task-3074645

Part-of: odoo/enterprise#36197
Co-authored-by: vava-odoo <vava@odoo.com>