[ADD] l10n_fr_reports : new IMB report
Add a new report, a variant of the Profit and Loss report, for the french localization.
(IMB) Intermediate management balances = (SIG) Soldes Intermédaires de gestion
In France, the structure of the P&L is useless management-wise.
This IMB report structures the information in a meaningful way that helps the user identify some ratios he is interested in.

closes odoo/enterprise#38553

Task: 3164564
X-original-commit: fddec479a98f6fe2996718d59efe4e4293164bb4
Signed-off-by: Quentin De Paoli <qdp@odoo.com>
Signed-off-by: hupo-odoo <hupo@odoo.com>