[ADD] l10n_fr_reports : new IMB report
Add a new report, a variant of the Profit and Loss report, for the french localization.
(IMB) Intermediate management balances = (SIG) Soldes Intermédaires de gestion
In France, the structure of the P&L is useless management-wise.
This IMB report structures the information in a meaningful way that helps the user identify some ratios he is interested in.

closes odoo/enterprise#37238

Task: 3164564
Signed-off-by: Quentin De Paoli <qdp@odoo.com>