[FIX] hr_payroll: todo_list use component Record
This commit aims to fix two bugs in the client action "hr_payroll_dashboard":
1. Opening the debug menu displays an error
2. Editing a note causes a crash

Why is this?
1. The useModel hook was created to be used only in a view. Using it
in a client action will cause the debug menu to think it is in a view,
which will cause the crash.

2. Using the html field in web_editor mode requires passing the setDirty
props to it. This will be used when the field is edited. In this case,
we use the FieldHtml component without this prop. So we will have a
crash when we edit the field.

Solution:
1. We will create our model and load it without using the useModel hook.

2. Just pass the setDirty prop to the html field.

closes odoo/enterprise#37214

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>