[FIX] l10n_cl_edi: add reception of services or goods
according to resolution 19.983.

Enable claim, acceptance and reception. Also in test mode.
Before these changes, the reception could be made from the SII site,
but now we can use a button to create this reception directly from Odoo.

Because there are multiple steps to be done now, the final state
is when we have both the reception of goods and the acceptance of the
invoice.

X-original-commit: 82230729814b727865e1de6843ab8f57434fce5b
Part-of: odoo/enterprise#43431