[FIX] sale_subscription: remove quick create from subscription kanban
before this commit, on clicking the quick create + button
from the subscription kanban view, is asking to input a
value in the quick create pop up which is currently
stored no where in the record.

initially when the create get called from name_create
the sequence is assigned for the code field and entered
name in the quick create pop up is passed to the name
field, when the user enter the further details like
customer, subscription template and hit save, the value
from the code which contains sequential value is saved
to the name field, thus we loose the value entered in
the quick create pop up.

* open subscription kanban
* click + button
* enter some values in the pop up and save
* enter further details and save
* entered value quick pop up is not stored anywhere in record

after this commit, the quick create option from the subscription kanban will be removed.

closes odoo/enterprise#41788

X-original-commit: 1a0df31b35c779a5f73f3be169521fa7a81b7f1a
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>