[FIX] knowledge: make name_get / name_search symmetrical
This commit aims to make the 'name_search' symmetrical to the name_get for the
knowledge.article model.

As we append the icon (emoji) before the article name, when searching based on
that same syntax '[emoji] name' we need to return the appropriate results.

This is especially important since some flows, such as exporting and
re-importing records, are based on name_get / name_search to match records
(for example when importing the article parent record, without this override it
will never match).

In addition, we add the "name" field into the list of default exported fields
for import compatible exports.

Task-3175039

closes odoo/enterprise#37617

X-original-commit: dc3ea160adf41cb8914116d6044c23ca969a48d2
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>