[IMP] knowledge: rework knowledge property panel UI
This commit adds a Helper inside the panel for the Properties inside the
corresponding panel.
With the addition of the helper we are removing the title from the panel and
centering the addProperty button when it is present.

We are also changing the behavior of the flow. When the user wants to add a
property to an article, that doesn't already contains one, he needs to actively
click on the addProperty button inside the panel.
It is no longer triggered when clicking on the button of the Topbar.

task-3096030

closes odoo/enterprise#38074

Related: odoo/upgrade#4584
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>