[FIX] sale_planning, project_timesheet_forecase_sale: fix demo data
Issue:
There is a difference between the `allocated_hours` and their
contribution to the "total" of the day in the planning demo data.

Steps to reproduce:
- Install Planning, Sale, Timesheet (with demo data)
- Go to Planning, filter for "Abigail Peterson" as resource
- Notice that for the planning, the allocated hours are 4h@100%, but
  the contribution for the "total" is 3h.

Reason for the problem:
Since version 16.0, `allocated_hours` are taking into consideration
the working calendar of the resource on the planning slot. The
`allocated_hours` and the `allocated_percentage` are co-dependant
computed fields, so when we set the `allocated_hours` we don't
modify the `allocated_percentage` (this is wanted behaviour). The
problem arises in the demo data where we set the `allocated_hours`
by hand, but the `allocated_percentage` should have been adjusted to
the correct %, taking into consideration the working calendar. For
example in some slots, we have for a shift of 7AM -> 11AM a manual
allocation of 4h, but the `allocated_percentage` is set to 100%,
instead of 133.33% it should've been.

Fix:
Remove the manual set of the `allocated_hours` in the demo data, so
it can be recomputed correctly, taking into account a default 100%
`allocated_percentage`. This way there shouldn't be any difference
between the `allocated_hours` on the slot and the contribution to
the "total".

Affected versions:
- 16.0
- saas-16.1
- master

opw-3184923

closes odoo/enterprise#37276

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>