[FIX] l10n_ec_edi: make report_withhold primary
Before this commit, the report_withhold in l10n_ec_edi inherited account's  tax_group_totals template without the primary attr set to True.
This meant it was replacing a condition in the original template and thus affecting other localizations.
This commit fixes that, thank you @BIB for pointing it out.

task-3011311

closes odoo/enterprise#37280

Signed-off-by: Josse Colpaert <jco@odoo.com>