[IMP] knowledge: do not set default title for articles
Currently, we use the value "Untitled" as default value for the title
field. When the user clicks on the article title, we will use the first
heading of the article as title if the current title is equivalent to
the translation of "Untitled". This comparison will obviously go wrong
if the users speak different languages.

Instead of using "Untitled" as default value, we will allow the article
title to be null and we will update the user interface to display
"Untitled" when the article does not have any title. The placeholder
"Untitled" will then be properly translated in the user's language
which was not the case when the title was stored in db.

task-3103594

closes odoo/enterprise#35061

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>