[IMP] knowledge: insert external embedded views in collaboration
Inserting Embedded views cannot happen from the backend anymore, since someone
could be currently editing the article which has to get the embedded view.

This commit introduces a way to insert an embedded view (or view link) using the
KnowledgeCommandsService, which will store the embedded view HTMLElement to be
inserted until the target record is accessed in Knowledge.

The HTMLElement will then be inserted like any other command from the editor,
and that change will be shared with collaborators if they are connected.

Since there is no way to know beforehand if the user will successfully connect
to the collaborative session, the content is inserted once when the user
accesses the record in Knowledge, and if and when his content is reset when a
collaborative session is joined, the content will be inserted again (since any
change would have been removed to get the current changes of the collaborator).

Task-2821374

Part-of: odoo/enterprise#33483