[FIX] l10n_ec_edi: remove L10nEcCertificate order
The order of that class was previously used for automatic selection of the most up-to-date certificate.
Since that selection is now manually done by the user that order is no longer necessary.
Moreover, the fields in the order are now computed fields which prevents them from being used for that purpose (warnings were being issued in the console).
Because of the reasons above, this commit removes the _order for the L10nEcCertificate class.

task-3011311

Part-of: odoo/enterprise#37280