[FIX] knowledge: allow moving an article that is not in the sidebar
When trying to use the "move to" feature on an article that is not shown in the
sidebar (typically because you can not see its parent), it would show a
traceback.

This is caused by trying to extract the "name" of the article from the sidebar
element, when we could instead simply use the record data.

Task-3062064

closes odoo/enterprise#39946

X-original-commit: bc249808d9a9fa4275b8e4df75e972b8f8cf6a01
Related: odoo/odoo#118929
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>