[FIX] approvals: apply approvers sequence
Steps to reproduce:
- configure a type of approval with no approvers,
approvers sequence enabled and a minimum approval
greater than or equal to 3;
- create an approval with this type;
- select 3 approvers;
- with the first approver, approve it.

Issue:
All other approvers have an activity created.
While the "approver sequence" parameter is enable.

Cause:
We only keep the next approver if he is defined in the type of approval.
We should keep only the next approver in all cases where the
"approver sequence" parameter is activated.

opw-3177434

closes odoo/enterprise#37370

X-original-commit: 73e27873d626a46bc61e27cf8c716bd75b874a52
Signed-off-by: Kevin Baptiste <kba@odoo.com>
Signed-off-by: Lefebvre Thomas (thle) <thle@odoo.com>