[IMP] sign: provide next documents to sign
The Thank You dialog now lists up to 3 most important items for the user to sign
and gives options to start signing them or ignore them.
Signing buttons in the dialog now land the user directly into signing mode.
A Download Document button now replaces the View Document button. It is only
visible if the document is completely signed.
This commit also includes various related improvements and fixes.

Notably, a lot of business logic has now been moved to the server, which also
makes the post-signing code more efficient. As this code is only used for this
one purpose, it makes more sense to do all processing on the server and return
data in the format we need.

task-3072737

closes odoo/enterprise#34649

Related: odoo/upgrade#4187
Signed-off-by: Arnaud Joset <arj@odoo.com>