[FIX] spreadsheet_dashboard_edition: prevent loss of user defined dashboards
ISSUE
=====
Currently, the dashboard data are in `noupdate=False`. This is a desired
behaviour as we want users to be able to benefit from the fixes to the
dashboards by updating the module.
However, since 9b13857, The downside is that the latter action will also
destroying the modifications of the client if they decided to edit the
dashboard to suit their needs.

FIX
===
We assume that once the user edits the dashboard, he will want to stick
with his version and doesn't want his work to be erased. We therefore
pass the dashboard external id to `noupdate=True` to preserve it.

If a user wants to benefit from a fixed dashboard even after modifying
it on their own, they can simply delete the record and update the
corresponding dashboard module.

OPW-3165458

closes odoo/enterprise#37320

Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>