[FIX] web_gantt: correct pill display name
When the attribute pill_label is set to true in a gantt view arch, some
computation of display names is done for pills. It turns out that this
computation was in many cases wrong. For example, it was possible that
a display name "" was returned or that some pill was incorrectly seen as
spanning over serveral days, weeks, or months. In this commit, we make
sure that the display names are correctly computed.

Part-of: odoo/enterprise#37091