[FIX] l10n_de_reports: datev export main account when same account
To reproduce:
With German company,create a Misc entry, with 2 lines on debit
with same account and 2 lines on credit.
Export DateV on general ledger
The lines appear without gegenkonto.

To fix:
We should look if we have a single account instead of looking if
we have a single line.

To note:
We still don't fix the case with 2 different accounts on credit and
2 different accounts on debit on a single Misc move.
This will be fixed in another task.

opw-3166295

closes odoo/enterprise#38099

X-original-commit: 9827b9e36804647459d7687f82e7a0454121f62c
Signed-off-by: Laurent Smet <las@odoo.com>
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>