[FIX] account_reports: fix tax report audit cell
In the case of the tax report, depending on the fiscal
postion's configuration, it can happen that we are
displaying a tax report from a different country
than the one from the current company.

This lead to incorrect domain computation when trying
to audit report line values.

With this fix, we are using the current report's country,
instead of the company's.

task-3199066

closes odoo/enterprise#37335

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>