[REF] l10n_cl_edi: remove test 110 (DTE generation for export)
Since the introduction of the `l10n_cl_edi_exports` module by
https://github.com/odoo/enterprise/pull/34031,
the generation of DTEs for exports is handled by that module.

That module provides a test for the DTE, and in fact disables the test
provided by `l10n_cl_edi`.

It makes sense to remove that test altogether, rather than just disable
it.

closes odoo/enterprise#37337

Signed-off-by: Josse Colpaert <jco@odoo.com>