[FIX] voip: always look for partner related to a call activity
Commit 5ba660f87a71f068ffb13f54f8de63d90ffcc166 changed the way
phonecalls are created from 'phonecall' activities. As a result, if an
activity is already associated with a phone number, the related partner
is never computed, resulting in this information remaining missing.

The goal of this commit is to restore the previous behavior, that is,
looking for the relevant partner even if a phone number is already known
for the given activity.

Task-3199796

closes odoo/enterprise#37352

Signed-off-by: Louis Wicket (wil) <wil@odoo.com>