[IMP] knowledge: reintroduce emojis in demo articles
This commit reverts the changes made in commit [1], in which the emojis
in xml files had been removed to avoid some crash that happened on some
MacOS with some specific lxml and python versions.

This commit also adds a pre_init_hook that will try to parse an emoji
and show a message advising to reinstall the lxml package without binaries
if parsing the emoji crashed.

Task-2993768

[1]: https://github.com/odoo/enterprise/commit/fd09b6f900e82d532505ad46905c549d1c237262

closes odoo/enterprise#37358

Signed-off-by: David Beguin (dbe) <dbe@odoo.com>