[FIX] currency_rate_live: make Bank of Peru exchange rate service work with non-English locales
## The issue
Currently, automatic currency rate updates with Bank of Peru do not work
if your locale is not English.

The Bank of Peru exchange rate service return dates in *English* with
the following format: `%d.%b.%y`.
Internally, we parse that date using`datetime.datetime.strptime`.

However, parsing fails when the locale is not English because `strptime`
uses the locale when parsing.
For example, trying to parse `29.Dec.22`with Spanish locale does not
work because `strptime` expects `29.Dic.22`.

## Fix
Replace the short month names with numbers.

opw-3091837

closes odoo/enterprise#37670

X-original-commit: d6d1bce48d0efbfa4bab9b320aaffe88377174de
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Nshimiyimana Serge Séna (sesn) <sesn@odoo.com>