[IMP] sale_subscription: Notify salesperson for message on invoice
Invoice in subscription are mostly generated by OdooBot, this
thus doesn't add the salesperson as a follower on the invoice.

We don't want the salesperson to be a follower on every invoice as
he would just be spammed by notification if everything goes
according to plan. However if the client post on the chatter
we still want the salesperson to be notified.

This PR add the salesperson as recipient of any non notification
written on the chatter, meaning that the salesperson receive a
notification or mail depending on it's notification preference set
in Odoo.

task-id : 3084135

closes odoo/enterprise#39116

X-original-commit: 4b3d2a7ed7bd99fa36f09631c2fd124ba5db5a28
Signed-off-by: Arnaud Joset <arj@odoo.com>
Signed-off-by: Damhaut Florian (flda) <flda@odoo.com>