[IMP] knowledge: change background buttons dark mode embeds
This commit removes the background of the buttons `Rename` and `Open` that are
specific to the embedded List and Kanban views for article items.
This way in dark mode the buttons doesn't stand out as much as before and have
a style similar to the buttons presents inside the embedded view by default
(filter, groupby, favorites).

task-3191812

closes odoo/enterprise#38538

Signed-off-by: David Beguin (dbe) <dbe@odoo.com>