[REF] helpdesk: refactor helpdesk and its report search view
Before this commit in-order to make both view same we add same content
(filter,group_by..) twice which was making it hard to keep track as lots
of search content coming from other bridge application and we have
redundant code.

So this commit make a base search view which contains all content of search
view and use it in both helpdesk and helpdesk report search views.

task-3186587

closes odoo/enterprise#37847

Related: odoo/upgrade#4405
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>