[FIX] web_studio: studio navbar to listen to the registry at setup
Be in studio, editing some view.
On the navbar, click on the button to go to the studio home menu.

Before this commit, the buttons "Edit Menu" and "New Model" were still present.
This was because the registry was not listened to at the right time.

After this commit, the buttons are removed as they should be if one switches from
editing a view to the studio home menu.

closes odoo/enterprise#37415

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>