[FIX] l10n_be_reports: fix partner VAT listing XML eport
Before this commit, trying to export the partner VAT listing
with load more active would generate a traceback.

This occured when trying to retrieve the VAT number from
the last report line since it is the "load more" line.

The additional problem was that the remaining partners were
simply omitted from the XML.

The fix consists in setting the `print_mode` to `True` in
the context before calling `_get_lines()` in order to ignore
the `load_more` limit.

task-3200219

closes odoo/enterprise#37486

X-original-commit: bbd68e105f2cdeb183239f570a66054ed8f636ba
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: John Laterre (jol) <jol@odoo.com>