[IMP] account_online_synchronization: Fetch missing bank statements
This commit will add the possibility to fetch missing bank statements from
a journal. It can happen that when a user configures a bank sync journal,
he actually needs to get transactions from a previous date (i.e. to
close the quarter) or some statements arrived later and the user wants
to have them.

From the journal form, the user can open a wizard and select a date. It
will be taken to a list view where he can select which statements he
wants to import. Some headers will be displayed to inform the user of
possible problems. The user also has a filter available to show the
pending statements (so he can be informed that if they're currently
missing, they will be imported when posted).
A custom list view was needed for the headers to be shown and it also
allows us to better manage the return from the view (going back to the
journal form view while maintaining the breadcrumbs).

task-3193313

closes odoo/enterprise#37570

Signed-off-by: Cedric Snauwaert <csn@odoo.com>