[FIX] knowledge: ensuring only one selected article
This commit fixes a bug where on a remote DB when you click on multiple
articles in the sidebar, multiple of them would appear as *selected*.

This means that they would appear as highlighted even though they are not open.

Before this commit we were only removing the styling from the previous article.
Now we are selecting all the articles that are active and their emojis and
removing their style before highlightng the opened article.
This disables the previously mentioned behavior.

task-3094192

closes odoo/enterprise#37575

X-original-commit: 94b19a48d2fe874d52ea6c8657e4c8b47779fa29
Signed-off-by: David Beguin (dbe) <dbe@odoo.com>