[FIX] website_knowledge: hide published shared article when not member.
Before this commit, as published shared article are technically accessible for
everyone, the articles were displayed in the shared section for internal users.

After this commit, published shared articles are filtered out when the current
user is not member of the article.

Also, since user_can_read has been reworked to include website_published
when website is installed, user_has_access need to be used for the shared
articles to be hidden for every share users if they are not member.

Task-3202849

closes odoo/enterprise#37457

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>