[FIX] sign: inactive users are not subscribed anymore
Sign runs tests with root user which is inactive. However inactive users
should never be added as followers, which is now fixed in community. This
commit updates tests accordingly.

Tests should probably user a real user, but trying to convert them lead
to other issues in tests which require some cleaning, and which is outside
the scope of this fix.

Task-3046371 (Mail: Better Language Support in Composer)

closes odoo/enterprise#37478

Related: odoo/odoo#113522
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>