[FIX] mrp_plm: persist ECO BoM
Before this commit, when the ECO's product was changed, the ECO's BoM
was always changed too for this product's first BoM, which is annoying
when the ECO already has a BoM for this product but not the first one.
This commit avoids that and changes the ECO's BoM when its product is
changed only when its BoM is for another product.

Part-of: odoo/enterprise#35969