[FIX] l10n_be_reports: correctly format VAT numbers in the VAT listing
If some of your cutomers VAT numbers contains '.' or ' ', when you'll
export your Belgian Partner VAT Listing in XML, you'll get an XML that
is not compliant with the Belgian administration requirements.

To avoid this issue, we sanitize the VAT numbers in the report to get
those numbers correctly formatted.

X-original-commit: 021730c3113339c7404c609501281a710a222612
Part-of: odoo/enterprise#37608