[FIX] l10n_be_reports: export all lines in XML format
When you export your Belgian Partner Listing  for VAT, you get a
traceback if you have more than 80 lines because it'll try to generate
the XML with the extra line named 'load more'. So we need to pass in the
context the 'print_mode' to True, this will avoid to add this extra line
'load more', and won't traceback anymore.

closes odoo/enterprise#37608

X-original-commit: 05930ae9e5910806e4a567c65e263669ddc89fca
Signed-off-by: Masereel Pierre <pim@odoo.com>