[IMP] sale_subscription: remove subscription stage, add subscription state
To simplify the subscription logic, this PR aims at removing the
subscription's stage model and anything that depends on it to replace
if by a simpler selection field called `subscription_state`.

This field will merge the usage of the follwing fields :
 -stage_id
 -stage_category
 -subscription_management
 -renew_state

For clarity sake's these fields are now represented by the 7
state of `subscription_state` :
 -'Quotation', Quotation for a new subscription
 -'Renewal Quotation', Renewal Quotation for existing subscription
 -'In Progress', Active Subscription or confirmed renewal for active subscription
 -'Paused', Active subscription with paused invoicing
 -'Renewed', Active or ended subscription that has been renewed
 -'Churned', Closed or ended subscription
 -'Upsell', Quotation or SO upselling a subscription,
    upsells are not subscriptions

Following theses changes, massive overhaul have been made
to the different subscription views to enhance the
usability as well as the onboarding of this app.
One of the main focus of these change was to make
the subscription view more closely act like the sale's
view.

Additionally, the old `to_renew` behaviour has been removed
as it wasn't fitting the new plan for subscription and will
be replaced in the future.

task-id : 3190077

closes odoo/enterprise#37536

Related: odoo/upgrade#4386
Signed-off-by: Arnaud Joset <arj@odoo.com>