[FIX] knowledge: article form gets locked into readonly mode
When the "action_home_page" action of the knowledge.article opens an article
in readonly mode, it sets the "form_view_initial_mode" to readonly.

That results into locking the form view in readonly even when switching to an
article you have write access to.

That "form_view_initial_mode" is actually not necessary as the view will
correctly be set as readonly or not depending on the user access rights.

Task-3203549

closes odoo/enterprise#37528

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>