[IMP] hr_payroll: display country related calendars
Before this commit user was able to select all calendars
in structure type even if user is not allowed in that
calendar's company.

This commit restrict user to only show calendars that
are in structure type's county's company's calendar
and if county not selected display all calendars.

task-3078009

closes odoo/enterprise#37334

Related: odoo/odoo#109904
Signed-off-by: Kevin Baptiste <kba@odoo.com>