[FIX] hr_payroll_attendance: change context for the date parser
The date parser introduced with commit 9e677ea115c1494d029513428e3919f54d24b27f
does not allow parsing of some local dates (for example: 'févr.').

A solution is to change the language in the context with the code `en_US`.
(we are sure that this language is installed because it is the default).

Moreover, the language of the date format is not important in this case,
because it will be converted into a `datetime` object.

opw-3182713

closes odoo/enterprise#37829

X-original-commit: 4ab33d300006215146ddbf7a19e754f9470f0966
Signed-off-by: Kevin Baptiste <kba@odoo.com>