[FIX] account_taxcloud: singleton error when reconciling credit notes
In a taxcloud company, reconciliation can throw a singleton error if it
contains multiple credit notes using taxcloud fiscal position because
it searches for the reversed_entry based on all moves reconciled instead
of the current move in the loop.

This commit changes it to the current move in the loop.

opw-3202418

closes odoo/enterprise#37645

X-original-commit: 74d8517a178a20e3806e40d0ecaadcd32a099e23
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>
Signed-off-by: Beckers Thomas (tbs) <tbs@odoo.com>