[FIX] knowledge: adds dialog service topbar
This commit adds the service dialog to the new TopBar Component which was
forgotten when creating it.

task-3208255

[IMP] knowledge: adding a new test for the UI

This commit adds a new test to the UI tests of knowledge.
This test ensures that moving an article via the button Move To inside the
panel containing all the options available to the user.

We are testing to move a private article as a child of a Root Article.
This use case is simple and enables us to quickly test the dialog.

closes odoo/enterprise#37557

Signed-off-by: David Beguin (dbe) <dbe@odoo.com>