[FIX] mrp_workorder: Get the google slide for manufacturing operations with specific page step
Steps to reproduce:
1- Create a BOM for a product, define an Operation on the product
2- Link a Google Slides document to the Operation as the Operation Worksheet.
3- Define an instruction on the Operation
4- Set Step Document to Specific Page of Operation Worksheet and define a page.

Issue: The step will not load the document.

Cause: tablet.js passes the wrong value to let know the viewer is a google slide url. Also when the good
value is passed the viewer always show the first page. That's because the SlideViewer has the page set in the
setup() method so it never change despite we change the step in the same document.

opw-3165142

closes odoo/enterprise#37558

Related: odoo/odoo#113727
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>