[FIX] approvals: cannot have same approver multiple times on request
Current behaviour:
You are able to have to have multiple times the same person as an
approver to a request.

Expected behaviour:
Approvers should be unique per request, it doesn't make sense to
have multiple times the same person as approver on a request.

Steps to reproduce:
- Install Approvals
- Create a new request, add 2 lines of the same approver
- Save, no error is thrown, so we have 2 lines with the same name.

Reason for the problem:
There is nothing no constraint on the uniqueness of the approvers
per request.

Fix:
Add an `@api.constrains` on the `approver_ids` of the request.

Affected versions:
- 14.0
- 15.0
- saas-15.2
- 16.0
- saas-16.1
- master

opw-3175530

closes odoo/enterprise#38439

X-original-commit: d4d94e8f4fe575dfcc96076fee8c59fbdcb48c79
Signed-off-by: Piryns Victor (pivi) <pivi@odoo.com>
Signed-off-by: Kevin Baptiste <kba@odoo.com>