[FIX] account_intrastat: extended options filter
A filter named options:'extended' or 'standard' exists on the intrastat
report to allow the user to include or exclude (respectively) the
transport and incoterm columns on the report.

This filter hasn't been functioning since v16.0 although the code was
written to adapt the options based on the setting of this filter, the
option are not adapted after the 'new_columns' are determined in the
code.

This code updates the options dictionary so that the columns in the
options reflect those determined by the filter.

closes odoo/enterprise#37724

X-original-commit: 9e5e13c781a18600c84a8d216a06ae22b3af3f84
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Daniel Kosky (dako) <dako@odoo.com>