[FIX] mrp: review and tablet view behavior correction
Review corrections

The tablet behavior was uncorrect.
It was previously possible to bypass the login and to work on the workorder.
The blocking screen was no working anymore.

Now, when entering the tablet view, we check if there is an admin of the session (if not => blocking screen and login)
The name on the top right of the tablet view is the admin of the session and the avatars are the people currently timesheeting on it.
To change admin, the user needs to click on his avatar/name in the selection popup.

Note that anyone can pause/restart a employee timesheet on this view

closes odoo/enterprise#34786

Related: odoo/odoo#107473
Signed-off-by: William Henrotin (whe) <whe@odoo.com>