[IMP] account_batch_payment, account_sepa: view improvements
When clicking on "X Payment(s) to send via SEPA" on the accounting
dashboard (on a bank journal), the view that was opened was a reduced
version of the one accessed via Vendors > Payments, hiding the
"Create Batch" button and some columns.

This commit opens the same "full" view via the link on the dashboard.
It filters the view on outgoing payments to send.
It also adds a space between the number and "Payments to send ...".

Finally, to be able to quickly see the status of a batch payment in the
payments view, we add the batch payment status in its name retrieved in
the list view.

task-3205654

closes odoo/enterprise#39175

X-original-commit: b4bebabd5dc49a435bc2feed863d177716bcb697
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Dylan Kiss (dyki) <dyki@odoo.com>