[FIX] website_knowledge: blocks click on the added div
This commit fixes a bug where a user with only read access on an article could
try to publish the article on the frontend but this would lead to a crash based
on how the click event was implemented for the div.
The event would click on the toggle for the field, but this field is not
present on the view if the user only has read access to the article, so it
would try to click on an element that is no longer present in the DOM which led
to the crash of the application.

Now the event is checking if the button is present in the DOM before trying to
click on it. This solution is the simplest and the one that requires the less
changes to be effective as a fix, that is why we chose to do this.

task-3208630

closes odoo/enterprise#37576

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>