[FIX] website_knowledge: fix crash of website_knowledge widget
This commit fixes a bug where in readonly mode when clicking on the div
containing the button to publish an article on the frontend this would cause
the application to crash because the user cannot change the published state
of an article if he cannot edit it.

Now we are using a custom function patched to the PermissionPanel Component
that checks if the user can write before trying to publish the article, leading
to no more crashes.

task-3208630

closes odoo/enterprise#41742

X-original-commit: 1075646fdd30db0af1755d9380307e4682c4d29b
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>