[FIX] sale_timesheet_enterprise: add new timesheet instead of raising user error
Before this commit starting timer on invoiced timesheet raise user
error and because of that user can't able to stop timer or start
timer on other timesheets.

So this commit add new timesheet on timer stop instead of tring to
update invoiced timesheet this way user can add start and stop
timer on invoiced timesheets.

task-3205260

closes odoo/enterprise#43047

X-original-commit: 9d8b749ac449f18e1eae1a07b6cf0f3f8d48cd99
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>